move functions to lib and add caisteal shouts
