+++ lib/darkages.ts
