+++ lib/discord.ts
